Move automatic phase advancing to the client side.
Automatic phase advancing on the server side is too complicated
and leads to a bunch of bugs.  So do it on the client, except
for dead players, who we need to skip on the server because they
may no longer have a client connected.

This means we'll always go through StartSplitPhase, so we no
longer need StartMovePhase.