Be rigorous about post-battle summoning and reinforcing.
Split Game._end_battle into _end_battle1 and _end_battle2.

_end_battle1 sets pending_reinforcement or pending_summon as
appropriate, and players cannot advance the phase until the
summon or reinforcement is dealt with.

Add DoNotSummon and DoNotReinforce actions and methods so that
if a player cannot or chooses not to summon or reinforce, the
game can continue.

_end_battle2 does the actual removal of legions and scoring.

The pending_acquire action is not respected yet, so acquiring
is still broken.