Add a non-random strike method.
First do the strikers with only one target, as before.  Then compute
the total damage we expect to be able to do to the remaining targets
if all possible strikers hit them, and hit the biggest one that we
expect to be able to kill.  If we can't kill any, then go after the
one that we can do the most damage to.  Strike with the least
valuable striker first, so that the bigger strikers can defer their
decisions until they have more information.