Only process Player.die on the server side.
Add Game.master, to tell which Game instance is the server one.
The problem was out-of-sync scores (and thus acquires) between the
server and clients, based on incorrect half-points from unknown
legions.  This should be fixed by only computing half-points on the
server where the legions are known, and then passing them down using
Action.AddPoints.

This change makes me sad, because we've made it this far with Game
being identical on the server and client, but I don't see a better way.