Use battle_active_player_name not active_player_name during battle.
We sometimes get a Flee/Concede/AddPoints while the local Game thinks
the battle is still running.  In this case the AI sometimes needs to move
to keep the game from hanging, but it's the battle_active_player (based
on battle phase), not the active_player (based on game turn) who needs
to move.  Though if we're out of sync there are no real guarantees.