Merge AddPoints and info from RevealLegion into EliminatePlayer.
There was a timing error where sometimes the AddPoints would arrive
on the client before Legion.die was called, which could cause
disagreements about whether the legion could acquire an angel.
Combining more into a single action might help with that.

We still send the RevealLegion action for each of the dead player's
legions, in addition to putting the information in the EliminatePlayer
action, to avoid having to special-case too much.