Update results to support multiple AIs of the same type.  Refs #172
Previously both human and AI players were scored by name, which is fine
for humans but silly for AIs.  Now both humans and AIs are scored by
type, where type is the same as name for humans, but is basically class
(currently always CleverBot) plus BotParams for AIs.

The Trueskill algorithm doesn't allow the same player to play more than
one position in the same game, so some hacks were necessary for games
where more than one AI of the same type is playing.  If all players are
the same type, the game is not scored at all.  If some players are the
same type, then their mu is averaged and the maximum sigma is taken.