Fix Undo/Redo split, by using just SplitLegion not RevealLegion.
Using a common SplitLegion (without creature names) then
two RevealLegion actions to inform the owner of the content of
the legions didn't undo correctly, because it was three actions
that logically did the work of one action.

So we're back to two SplitLegion actions, one with actual creature
names for the owner, and one without for everyone else.  We filter
the censored version in the owner's User, to avoid problems from
near-duplicate actions.