Re-add --min-players and --max-players to stresstest
It's usually good to randomly vary the number of players when training
the AI, because some tactics work better for 2-player games and others
work better for 6-player games.  But sometimes I want to force all the
games to 6 players, like when I was hunting for a file descriptor leak
and wanted it to happen faster.