Make AIClient exit if it fails to join a game.
There's a race in initialization that sometimes led to extra AIClients
being created, failing to login (because the game already has
max_players players), and then hanging around forever consuming extra
file descriptors on the server.  This eventually led to the server
running out of file descriptors.  It would be better to totally fix the
race, but this makes at least it much less harmful by making the excess
AIClients go away quickly.